+++ tests/tasks/test_workflows_cellpose_segmentation.py
@@ -690,7 +690,18 @@ def test_workflow_with_per_FOV_labeling_via_script(
     res = subprocess.run(shlex.split(command), **run_options)  # type: ignore
     print(res.stdout)
     print(res.stderr)
-    error_msg = f"ERROR model_type={INVALID_MODEL_TYPE} is not allowed"
+    # If this check fails after updating the cellpose version, you'll likely
+    # need to update the manifest to include a changed set of available models
+    # See https://github.com/fractal-analytics-platform/fractal-tasks-core/issues/401 # noqa E501
+    error_msg = (
+        "unexpected value; permitted: 'cyto', 'nuclei', "
+        "'tissuenet', 'livecell', 'cyto2', 'general', 'CP', 'CPx', "
+        "'TN1', 'TN2', 'TN3', 'LC1', 'LC2', 'LC3', 'LC4' "
+        f"(type=value_error.const; given={INVALID_MODEL_TYPE}; "
+        "permitted=('cyto', 'nuclei', 'tissuenet', 'livecell', "
+        "'cyto2', 'general', 'CP', 'CPx', 'TN1', 'TN2', 'TN3', "
+        "'LC1', 'LC2', 'LC3', 'LC4'))"
+    )
     assert error_msg in res.stderr
     assert "urllib.error.HTTPError" not in res.stdout
     assert "urllib.error.HTTPError" not in res.stderr